move test files for scopes into own sub directory
