+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/scopes/example1/views/ScopedFxmlViewC.java
@@ -1,4 +1,4 @@
-package de.saxsys.mvvmfx.scopes;
+package de.saxsys.mvvmfx.scopes.example1.views;
 
 import de.saxsys.mvvmfx.Context;
 import de.saxsys.mvvmfx.FluentViewLoader;
@@ -8,27 +8,29 @@ import de.saxsys.mvvmfx.InjectViewModel;
 import de.saxsys.mvvmfx.ViewTuple;
 import de.saxsys.mvvmfx.internal.viewloader.example.TestScope1;
 import de.saxsys.mvvmfx.internal.viewloader.example.TestScope2;
+import de.saxsys.mvvmfx.scopes.example1.Example1Scope1;
+import de.saxsys.mvvmfx.scopes.example1.Example1Scope2;
 import javafx.fxml.FXML;
 import javafx.scene.layout.VBox;
 
 public class ScopedFxmlViewC implements FxmlView<ScopedViewModelC> {
 
     @InjectViewModel
-    ScopedViewModelC viewModel;
+    public ScopedViewModelC viewModel;
 
     @FXML
-    VBox root;
+    public VBox root;
 
     @InjectContext
-    Context context;
+    public Context context;
 
-    ScopedFxmlViewD subViewDController;
-    ScopedFxmlViewD subViewDWithoutContextController;
+    public ScopedFxmlViewD subViewDController;
+    public ScopedFxmlViewD subViewDWithoutContextController;
 
     public void initialize() {
         ViewTuple<ScopedFxmlViewD, ScopedViewModelD> load = FluentViewLoader.fxmlView(ScopedFxmlViewD.class)
                 .context(context)
-                .providedScopes(new TestScope2())
+                .providedScopes(new Example1Scope2())
                 .load();
         root.getChildren().add(load.getView());
         subViewDController = load.getCodeBehind();
@@ -42,7 +44,7 @@ public class ScopedFxmlViewC implements FxmlView<ScopedViewModelC> {
 
     public void loadCorrectScopedView() {
         ViewTuple<ScopedFxmlViewD, ScopedViewModelD> load2 = FluentViewLoader.fxmlView(ScopedFxmlViewD.class)
-                .providedScopes(new TestScope1(), new TestScope2())
+                .providedScopes(new Example1Scope1(), new Example1Scope2())
                 .load();
         root.getChildren().add(load2.getView());
         subViewDWithoutContextController = load2.getCodeBehind();
