+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/scopes/example1/views/ScopedFxmlViewD.java
@@ -1,9 +1,8 @@
-package de.saxsys.mvvmfx.scopes;
+package de.saxsys.mvvmfx.scopes.example1.views;
 
 import de.saxsys.mvvmfx.FxmlView;
 import de.saxsys.mvvmfx.InjectViewModel;
 import javafx.fxml.FXML;
-import javafx.scene.layout.VBox;
 
 public class ScopedFxmlViewD implements FxmlView<ScopedViewModelD> {
 
@@ -14,6 +13,6 @@ public class ScopedFxmlViewD implements FxmlView<ScopedViewModelD> {
     public ScopedFxmlViewG subviewGController;
 
     @InjectViewModel
-    ScopedViewModelD viewModel;
+    public ScopedViewModelD viewModel;
 
 }
