+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/scopes/example1/views/ScopedViewModelC.java
@@ -1,12 +1,12 @@
-package de.saxsys.mvvmfx.scopes;
+package de.saxsys.mvvmfx.scopes.example1.views;
 
 import de.saxsys.mvvmfx.InjectScope;
 import de.saxsys.mvvmfx.ViewModel;
-import de.saxsys.mvvmfx.internal.viewloader.example.TestScope1;
+import de.saxsys.mvvmfx.scopes.example1.Example1Scope1;
 
 public class ScopedViewModelC implements ViewModel {
 
     @InjectScope
-    public TestScope1 injectedScope1;
+    public Example1Scope1 injectedScope1;
 
 }
