+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/scopes/example1/views/ScopedViewModelE.java
@@ -1,15 +1,18 @@
-package de.saxsys.mvvmfx.scopes;
+package de.saxsys.mvvmfx.scopes.example1.views;
 
 import de.saxsys.mvvmfx.InjectScope;
+import de.saxsys.mvvmfx.ScopeProvider;
 import de.saxsys.mvvmfx.ViewModel;
+import de.saxsys.mvvmfx.scopes.example1.Example1Scope3;
 
-public class ScopedViewModelG implements ViewModel{
+@ScopeProvider(scopes = {Example1Scope3.class})
+public class ScopedViewModelE implements ViewModel{
 
 	@InjectScope
-	public TestScope3 testScope3;
+	public Example1Scope3 testScope3;
 
 
-	public ScopedViewModelG() {
+	public ScopedViewModelE() {
 		System.out.println("new " + this.getClass().getSimpleName() + "()");
 	}
 
@@ -17,5 +20,4 @@ public class ScopedViewModelG implements ViewModel{
 		System.out.println(this.getClass().getSimpleName() + ".initialize()");
 		System.out.println("scope:" + System.identityHashCode(testScope3));
 	}
-
 }
