+++ mvvmfx/src/test/java/de/saxsys/mvvmfx/scopes/example1/views/ScopedViewModelG.java
@@ -1,14 +1,16 @@
-package de.saxsys.mvvmfx.scopes;
+package de.saxsys.mvvmfx.scopes.example1.views;
 
 import de.saxsys.mvvmfx.InjectScope;
 import de.saxsys.mvvmfx.ViewModel;
+import de.saxsys.mvvmfx.scopes.example1.Example1Scope3;
 
-public class ScopedViewModelF implements ViewModel{
+public class ScopedViewModelG implements ViewModel{
 
 	@InjectScope
-	public TestScope3 testScope3;
+	public Example1Scope3 testScope3;
 
-	public ScopedViewModelF() {
+
+	public ScopedViewModelG() {
 		System.out.println("new " + this.getClass().getSimpleName() + "()");
 	}
 
@@ -16,4 +18,5 @@ public class ScopedViewModelF implements ViewModel{
 		System.out.println(this.getClass().getSimpleName() + ".initialize()");
 		System.out.println("scope:" + System.identityHashCode(testScope3));
 	}
+
 }
