--- mvvmfx/src/test/resources/de/saxsys/mvvmfx/scopes/ScopedFxmlViewD.fxml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.layout.VBox?>
-<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.mvvmfx.scopes.ScopedFxmlViewD">
-
-</VBox>
