+++ mvvmfx/src/test/resources/de/saxsys/mvvmfx/scopes/example1/views/ScopedFxmlViewA.fxml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.layout.VBox?>
-<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.mvvmfx.scopes.ScopedFxmlViewB">
-
+<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.mvvmfx.scopes.example1.views.ScopedFxmlViewA">
 <children>
         <fx:include fx:id="subviewC" source="ScopedFxmlViewC.fxml"/>
         <fx:include fx:id="subviewE" source="ScopedFxmlViewE.fxml"/>
 </children>
-
 </VBox>
