+++ mvvmfx/src/test/resources/de/saxsys/mvvmfx/scopes/example1/views/ScopedFxmlViewB.fxml
@@ -1,9 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.layout.VBox?>
-<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.mvvmfx.scopes.ScopedFxmlViewA">
+<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.mvvmfx.scopes.example1.views.ScopedFxmlViewB">
+
 <children>
         <fx:include fx:id="subviewC" source="ScopedFxmlViewC.fxml"/>
         <fx:include fx:id="subviewE" source="ScopedFxmlViewE.fxml"/>
 </children>
+
 </VBox>
