+++ mvvmfx/src/test/resources/de/saxsys/mvvmfx/scopes/example1/views/ScopedFxmlViewC.fxml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.VBox?>
+<VBox xmlns:fx="http://javafx.com/fxml" fx:controller="de.saxsys.mvvmfx.scopes.example1.views.ScopedFxmlViewC" fx:id="root">
+
+</VBox>
